requirements.txt
absl-py==0.6.1
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.0.2
bs4==0.0.1
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dialogflow==0.5.1
Django==2.1.4
entrypoints==0.2.3
future==0.17.1
gast==0.2.0
google-api-core==1.6.0
google-auth==1.6.1
googleapis-common-protos==1.5.5
grpcio==1.16.1
h5py==2.8.0
idna==2.7
imgurpython==1.1.7
imutils==0.5.2
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
line-bot-sdk==1.8.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.3.0
pymongo==3.7.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.20.1
rsa==4.0
scikit-learn==0.20.1
scipy==1.1.0
Send2Trash==1.5.0
six==1.11.0
sklearn==0.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
